JNI: Use GetByteArrayRegion/SetByteArrayRegion (#132)
JNI: Use GetByteArrayRegion/SetByteArrayRegion

We used GetByteArrayElements/ReleaseByteArrayElements which would "pin" the JVM object and give us a pointer to its memory that we could use in native calls.
This would interact with the GC and cause issues that would be platform-specific and very hard to debug.

We now use GetByteArrayRegion/SetByteArrayRegion which do not "pin" the JVM ovject at all. Instead we copy data to a local buffer and use it, or copy it
back from a local buffer to a JVM array.