Fix android tests (#136)
* Fix android tests

They had been silently ignored on our CI.
Test resources are duplicated in src/androidInstrumentedTest/assets and src/commonTest/resources for now until we find a way to share them.

* Upgrade gradle to 8.14.3 and android gradle plugin to 8.12.3

This fixes `WARNING: D8: Unexpected error during rewriting of Kotlin metadata for class` when building the android library.

* Update build after gradle upgrade

* Upgrade dokka plugin

* Fix potential memory leak when some methods fail

There was an off-by-one error in how we freed dynamically allocated memory in case of failure.
This only happens in methods that take arrays of pubkeys/nonces/partial signatures, and only when they fail.

* Format code with clang-format (no functional changes)

* Fix compilation warnings (no functional changes)

* Use kotlin 2.2.21

* Use AGP 8.13.1

* Android CI tests: set emulator disk size to 5G

* Use NDK 28.2.13676358 (same as Github CI)